Some adjustments for example

+++ embassy-stm32/src/fmac/mod.rs
@@ -37,13 +37,12 @@ trait SealedInstance {
         self.write_input(x.inner as u16);
     }
 
-    //fn rdata() -> *const u32 {
-    //    Self::regs().rdata().as_ptr() as *const u32
+    //fn rdata() -> *const u16 {
+    //    Self::regs().rdata().as_ptr() as *const u16
     //}
 
-    #[allow(dead_code)]
-    fn wdata() -> *mut u32 {
-        Self::regs().wdata().as_ptr() as *mut u32
+    fn wdata() -> *mut u16 {
+        Self::regs().wdata().as_ptr() as *mut u16
     }
 }
